third_party/blink/web_tests/fast/dom/shadow/selections-in-shadow-expected.txt

CONSOLE WARNING: line 40: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.



CONTAINER1 --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1 --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2 --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> CONTAINER1
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> CONTAINER2
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> IN HOST
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> SPAN IN NESTED IN SHADOW OLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> SPAN 1 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> SPAN 2 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER --> SPAN 3 INOLDER
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER1: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
CONTAINER2: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
IN HOST: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN IN NESTED IN SHADOW OLDER: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 1 INOLDER: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 2 INOLDER: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)
SPAN 3 INOLDER: double click
PASS internals.treeScopeRootNode(window.getSelection().anchorNode) is internals.treeScopeRootNode(window.getSelection().focusNode)





PASS successfullyParsed is true

TEST COMPLETE
